refactor: move interface into model
